requirements.txt
Django==1.5
Jinja2==2.6
MySQL-python==1.2.4
Pygments==1.6
South==0.7.6
Sphinx==1.1.3
Werkzeug==0.8.3
argparse==1.2.1
coverage==3.6
cssselect==0.8
django-autocomplete-light==1.1.25
django-codemirror-widget==0.3.0
django-debug-toolbar==0.9.4
django-extensions==1.0.3
django-mptt==0.5.5
django-selectable==0.7.0
django-suit==0.2.1
django-tinymce==1.5.1b4
docutils==0.10
fuzzywuzzy==0.1
ipdb==0.7
ipython==0.13.1
lettuce==0.2.15
https://pypi.python.org/packages/source/l/lxml/lxml-3.2.0.tar.gz
mechanize==0.2.5
nose==1.2.1
openpyxl==1.6.1
pyquery==1.2.4
pytz==2013b
pyzmq==13.0.2
readline==6.2.4.1
salad==0.4.14
selenium==2.31.0
splinter==0.5.0
sure==1.2.0
tornado==3.0.1
wadofstuff-django-serializers==1.1.0
xlrd==0.9.0
zope.event==4.0.2
zope.schema==4.3.2
zope.testbrowser==4.0.2
pygeocoder==1.1.4
django-localflavor-us==1.1
django-apptemplates==0.0.1
django-admin-sortable
django-suit-redactor
django-suit-ckeditor
django-admin-bootstrapped
django-reversion
django-select2
django-page-cms
django-staticfiles
django-authority
django-forms-builder==0.9
django-sql-explorer==0.2